requirements.txt
openai==1.14.1
tiktoken==0.7.0
python-dotenv==1.0.1
